don't comapare so that -1 is more than INT_MAX

+++ src/socket.c
@@ -570,7 +570,7 @@ static ssize_t ssh_socket_unbuffered_write(ssh_socket s,
 #else
     s->last_errno = errno;
 #endif
-	if (w < len) // shouldn't assume it will block unless we couldn't send it all
+	if (w < (int)len) // shouldn't assume it will block unless we couldn't send it all
 		s->write_wontblock = 0;
     /* Reactive the POLLOUT detector in the poll multiplexer system */
     if (s->poll_handle) {
